mcp-langgraph-rag/pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "mcp-langgraph-rag"
version = "0.1.0"
description = "LangGraph MCP server with local FAISS-backed RAG tools."
readme = "README.md"
authors = [{ name = "Your Team" }]
license = { text = "MIT" }
requires-python = ">=3.13"
dependencies = [
  "fastmcp>=0.1.4",
  "langgraph>=0.1.0",
  "langchain>=0.2.15",
  "langchain-community>=0.2.8",
  "langchain-text-splitters>=0.2.2",
  "langchain-openai>=0.1.22",
  "faiss-cpu>=1.8.0",
  "pydantic>=2.8.2",
  "python-dotenv>=1.0.1",
  "typer>=0.12.5",
  "rich>=13.7.1",
]

[project.optional-dependencies]
dev = [
  "black>=24.8.0",
  "mypy>=1.11.1",
  "pytest>=8.3.2",
  "pytest-asyncio>=0.23.7",
  "ruff>=0.6.1",
  "pre-commit>=3.8.0",
  "tiktoken>=0.7.0",
]

[project.scripts]
ingest = "scripts.build_index:main"
demo = "scripts.demo_queries:main"

[tool.hatch.build.targets.wheel]
packages = ["app", "scripts"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "ASYNC"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.13"
packages = ["app", "scripts"]
strict = true
ignore_missing_imports = true

[tool.coverage.run]
branch = true
source = ["app"]
